refactor: move to lib crate
