+++ src/bin/main.rs
@@ -1,11 +1,8 @@
 use anyhow::Ok;
-use repository::Repository;
 use tower_http::trace::TraceLayer;
 use tracing_subscriber::EnvFilter;
 
-mod handler;
-mod repository;
-mod utils;
+use trao_judge_backend as lib;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
@@ -13,10 +10,10 @@ async fn main() -> anyhow::Result<()> {
         .with_env_filter(EnvFilter::try_from_default_env().unwrap_or("info".into()))
         .init();
 
-    let app_state = Repository::connect().await?;
+    let app_state = lib::Repository::connect().await?;
     app_state.migrate().await?;
 
-    let app = handler::make_router(app_state).layer(TraceLayer::new_for_http());
+    let app = lib::make_router(app_state).layer(TraceLayer::new_for_http());
 
     let listener = tokio::net::TcpListener::bind("0.0.0.0:3000").await?;
